Update Syllabus.md
Anyone who has read him will hate me, but I added Floridi from the new Texts and Technology list. I don't think that people would have to read all of him, but I think it wouldbe important to get people thinking about the ethics of internet research early in the program and to not make thinking about ethics an elective.  I am open to changing this, but it is a suggestion.